refactor(monitoring): reorganize and enhance monitoring components and services
- Updated the AdminMonitoringSection to utilize shared components for better modularity.
- Refactored monitoring service to streamline type exports and improve code organization.
- Enhanced documentation for monitoring features, including detailed descriptions of new components and their functionalities.
- Improved the structure of monitoring-related files for better maintainability and clarity.
