+++ src/components/admin/monitoring/stats/index.ts
@@ -1,0 +1,7 @@
+// Re-export all stats components for convenient importing
+export { DatabaseStats } from "./DatabaseStats";
+export { AuthStats } from "./AuthStats";
+export { EdgeFunctionStats } from "./EdgeFunctionStats";
+export { EmailStatsDisplay } from "./EmailStatsDisplay";
+export { CloudflareStatsDisplay } from "./CloudflareStatsDisplay";
+export { VercelStatsDisplay } from "./VercelStatsDisplay";
